requirements.txt
django==4.2.8
python-dotenv==1.0.0
djangorestframework==3.14.0
Markdown==3.5.1
django-filter==23.5
djangorestframework-simplejwt==5.3.1
dj-rest-auth==5.0.2
dj-rest-auth[with_social]
django-cors-headers==4.3.1
drf-yasg==1.21.7
drf-yasg[validation]